# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hệ thống Chấm công QR API
  description: |
    API for QR Code Attendance System with GPS verification.

    ## Authentication
    All endpoints require JWT Bearer token except `/auth/login`.

    ## Rate Limiting
    - Authentication: 5 requests/minute
    - Attendance: 10 requests/minute
    - General: 100 requests/minute
  version: 1.0.0
  contact:
    name: API Support
    email: support@company.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development
  - url: https://attendance.company.com/api/v1
    description: Production
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get current user info
      description: Get authenticated user details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /attendance:
    get:
      tags: [Attendance]
      operationId: getAllAttendance
      summary: Get all attendance records
      description: Get all attendance records (Admin/HR only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: userId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: type
          in: query
          schema:
            type: string
            enum: [IN, OUT]
      responses:
        '200':
          description: Attendance records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendancePageResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /attendance/export:
    get:
      tags: [Attendance]
      operationId: exportAttendance
      summary: Export attendance to CSV
      description: Export attendance records to CSV file (Admin/HR only)
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Get user by ID
      description: Get user details (Admin or self)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      operationId: updateUser
      summary: Update user
      description: Update user information (Admin or self)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete user
      description: Delete user account (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: User deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /config/gps:
    get:
      tags: [Config]
      operationId: getGPSConfig
      summary: Get GPS configuration
      description: Get GPS settings (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: GPS configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GPSConfigResponse'

    put:
      tags: [Config]
      operationId: updateGPSConfig
      summary: Update GPS configuration
      description: Update GPS settings (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GPSConfigRequest'
      responses:
        '200':
          description: GPS config updated
  /config/qr:
    get:
      tags: [Config]
      operationId: getQRConfig
      summary: Get QR code configuration
      description: Get QR expiration settings (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: QR configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRConfigResponse'

    put:
      tags: [Config]
      operationId: updateQRConfig
      summary: Update QR code configuration
      description: Update QR expiration settings (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QRConfigRequest'
      responses:
        '200':
          description: QR config updated
  /config/attendance:
    get:
      tags: [Config]
      operationId: getAttendanceConfig
      summary: Get attendance configuration
      description: Get attendance settings like late threshold (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Attendance configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceConfigResponse'

    put:
      tags: [Config]
      operationId: updateAttendanceConfig
      summary: Update attendance configuration
      description: Update attendance settings like late threshold (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceConfigRequest'
      responses:
        '200':
          description: Attendance config updated

  # ========== Leave Request Endpoints ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            expiresIn:
              type: integer
            user:
              $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [EMPLOYEE, ADMIN, HR_MANAGER]
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        shiftId:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
...
    UserUpdateRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        shiftId:
          type: integer
        status:
          type: string
...
    Attendance:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        timestamp:
          type: string
          format: date-time
        distance:
          type: number
        status:
          type: string
        flag:
          type: string
    AttendancePageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Attendance'
            page:
              type: integer
            size:
              type: integer
            totalElements:
              type: integer
            totalPages:
              type: integer
            summary:
              $ref: '#/components/schemas/AttendanceSummary'
    AttendanceSummary:
      type: object
      properties:
        total:
          type: integer
        onTime:
          type: integer
        late:
          type: integer
        absent:
          type: integer
        excused:
          type: integer
...
    GPSConfigResponse:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        threshold:
          type: number
        wifiFallback:
          type: boolean
    GPSConfigRequest:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        threshold:
          type: number
        wifiFallback:
          type: boolean
    QRConfigResponse:
      type: object
      properties:
        qrExpirationSeconds:
          type: integer
          description: QR code expiration time in seconds
          example: 30
    QRConfigRequest:
      type: object
      required: [qrExpirationSeconds]
      properties:
        qrExpirationSeconds:
          type: integer
          minimum: 1
          description: QR code expiration time in seconds (must be > 0)
          example: 30
    AttendanceConfigResponse:
      type: object
      properties:
        lateThresholdMinutes:
          type: integer
          description: Maximum minutes late still considered on-time
          example: 5
    AttendanceConfigRequest:
      type: object
      required: [lateThresholdMinutes]
      properties:
        lateThresholdMinutes:
          type: integer
          minimum: 1
          description: Maximum minutes late still considered on-time (must be > 0)
          example: 5

    # ========== Salary Config New ==========
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...